requirements.txt
flask
cairosvg
Pillow
requests